+++ src/main/java/net/meander/subtlyd/mixin/client/renderer/entity/LivingEntityRendererMixin.java
@@ -35,11 +35,12 @@ public class LivingEntityRendererMixin<T extends LivingEntity, S extends LivingE
         if (state instanceof LivingEntityRenderStateAccessor stateAccessor) {
             if (entity instanceof ClimberAccessor climberAccessor) {
                 extractClimberState(entity, climberAccessor, state, stateAccessor, partialTicks);
-            } else if (entity.getVehicle() instanceof ClimberAccessor spider) {
-                extractClimberJockeyState(entity.getVehicle(), spider, state, stateAccessor, partialTicks);
+            } else if (entity.getVehicle() instanceof ClimberAccessor arthropod) {
+                extractClimberJockeyState(entity.getVehicle(), arthropod, state, stateAccessor, partialTicks);
             }
         }
     }
+
 
     /**
      * Determines the rotations that should be set up.
@@ -90,6 +91,7 @@ public class LivingEntityRendererMixin<T extends LivingEntity, S extends LivingE
 
         stateAccessor.setIsJockey(true);
         stateAccessor.setClimbProgress(progress);
+
         if (progress > climbProgressThreshold && nearestWall != null) {
             float oldYaw = stateAccessor.getClimbRotation();
             state.bodyRot = Mth.rotLerp(progress, state.bodyRot, climberAccessor.getRotation(partialTicks));
@@ -100,11 +102,11 @@ public class LivingEntityRendererMixin<T extends LivingEntity, S extends LivingE
     /**
      * Sets up visual rotations and offsets for climbing entities.
      *
-     * @param accessor Interface for spider render state information.
+     * @param accessor Interface for arthropod render state information.
      * @param poseStack The pose stack of the climbing entity.
-     * @param bBH The bounding box height of the entity. Used for relative translations.
+     * @param bBHeight The bounding box height of the entity. Used for relative translations.
      */
-    private void setupClimberRotations(LivingEntityRenderStateAccessor accessor, PoseStack poseStack, float bBH) {
+    private void setupClimberRotations(LivingEntityRenderStateAccessor accessor, PoseStack poseStack, float bBHeight) {
         float progress = accessor.getClimbProgress();
         float yOffset;
         float zOffset;
@@ -112,7 +114,7 @@ public class LivingEntityRendererMixin<T extends LivingEntity, S extends LivingE
         if (progress > climbProgressThreshold) {
             if (!accessor.isJockey()) {
                 yOffset = 0.2F * progress;
-                zOffset = (-1.06F * bBH + 0.2F) * progress; // Linear function to prevent cave spiders from clipping into walls.
+                zOffset = (-1.06F * bBHeight + 0.2F) * progress; // Linear function to prevent cave spiders from clipping into walls.
             } else {
                 yOffset = 0.5F * progress;
                 zOffset = -1.6F * progress;